+++ history.md
@@ -3,9 +3,9 @@
 - 实现多选的高度扩展设置项
 - switch false判断
 
-### 2.1.27-beta
-- FormView Foot设置项添加loading/disabled
-- 准备拆解FormView
+### 2.1.27
+- FormView拆分Item和Foot，优化相关class
+- FormView auto设置项添加loading/disabled
 
 ### 2.1.26
 - AutoIndex BUG修复，从0开始改为从1开始